Added endpoint to ask seeding to stop
When called, the hypercore will be dropped from the seeding set and connections to peers dropped.

Also, when adding a core that already is being tracked, we now update the last seen time for it, essentially keeping the seeding alive for longer. If the app requests it’s core be seeded each time it is opened, this should keep them alive, even if there are not changes to the content.